Truncate varint to 32-bit integer.
This reverts an unintentional change in CL 226346424 where on 32-bit platforms
varint was implicitly truncated in some cases but not the others.

gfe-relnote: n/a (only affects 32-bit platforms)
PiperOrigin-RevId: 229679285

+++ http2/hpack/decoder/hpack_entry_decoder.cc
@@ -192,7 +192,7 @@ DecodeStatus HpackEntryDecoder::Resume(DecodeBuffer* db,
 
 bool HpackEntryDecoder::DispatchOnType(HpackEntryDecoderListener* listener) {
   const HpackEntryType entry_type = entry_type_decoder_.entry_type();
-  const uint64_t varint = entry_type_decoder_.varint();
+  const uint32_t varint = static_cast<uint32_t>(entry_type_decoder_.varint());
   switch (entry_type) {
     case HpackEntryType::kIndexedHeader:
       // The entry consists solely of the entry type and varint. See:
